Haberler alanında repopulation işlemleri ve validation işlemleri yapıldı resim silindiği zaman aynı anda klasörden silinmesi ve güncelleme yapıldığı zamanda klasörden silinmesi gerçekleştirildi.

+++ panel/application/controllers/News.php
@@ -70,6 +70,8 @@ class News extends CI_Controller
       }
 
       $this->form_validation->set_rules("title", "Başlık", "required|trim");
+      $this->form_validation->set_rules("description", "Açıklama", "trim");
+
       $this->form_validation->set_message(
           array(
               "required" => "<strong>{field}</strong> alanı doldurulmalıdır."
@@ -185,77 +187,6 @@ class News extends CI_Controller
         $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index", $viewData);
     }
 
-    //düzenlenen veriyi veri tabanına kaydetmek
-    public function update1($id){
-        $this->load->library("form_validation");
-
-        $this->form_validation->set_rules("title", "Başlık", "required|trim");
-        $this->form_validation->set_message(
-
-            array(
-                "required" => "<strong>{field}</strong> alanı doldurulmalıdır."
-            )
-        );
-        $validate = $this->form_validation->run();
-
-        if ($validate){
-
-            $update = $this->news_model->update(
-                array(
-                    "id" => $id,
-                ),
-                array(
-                    "url"         => converToSEO($this->input->post("title")),
-                    "title"       => $this->input->post("title"),
-                    "description" => $this->input->post("description")
-                )
-            );
-
-            if($update){
-                $alert = array(
-                    "title"   => "Tebrikler",
-                    "text"    => "İşleminiz başarılı bir şekilde gerçekleştirildi.",
-                    "type"    => "success"
-                );
-
-            }
-            else{
-                $alert = array(
-                    "title"   => "İşlem başarısız",
-                    "text"    => "Lütfen zorunlu olan alanları doldurunuz!",
-                    "type"    => "error"
-                );
-            }
-            $this->session->set_flashdata("alert", $alert);
-            redirect(base_url("news"));
-
-        }
-        else{
-            $viewData = new stdClass();
-
-            $item = $this->news_model->get(
-                array(
-                    "id" => $id
-                )
-            );
-
-            $viewData-> viewFolder = $this->viewFolder;
-            $viewData->subViewFolder = "update";
-            $viewData->form_error = "true";
-            $viewData->item = $item;
-
-
-            $alert = array(
-                "title"   => "İşlem başarısız",
-                "text"    => "Lütfen zorunlu olan alanları doldurunuz!",
-                "type"    => "error"
-            );
-            $this->session->set_flashdata("alert", $alert);
-            $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index", $viewData);
-            unset($_SESSION['alert']);
-        }
-    }
-
     public function update($id){
         $this->load->library("form_validation");
 
@@ -266,6 +197,7 @@ class News extends CI_Controller
         }
 
         $this->form_validation->set_rules("title", "Başlık", "required|trim");
+        $this->form_validation->set_rules("description", "Açıklama", "trim");
 
         $this->form_validation->set_message(
             array(
@@ -277,35 +209,67 @@ class News extends CI_Controller
 
         if ($validate){
             if($news_type === "image"){
-
                 if ($_FILES["img_url"]["name"] !== ""){
-                    $file_name = converToSEO(pathinfo($_FILES["img_url"]["name"], PATHINFO_FILENAME)). "." . pathinfo($_FILES["img_url"]["name"], PATHINFO_EXTENSION);
-                    $config["allowed_types"] = "jpg|jpeg|png";
-                    $config["upload_path"] = "uploads/$this->viewFolder/";
-                    $config["file_name"] = $file_name;
+                    $image = $this->news_model->get(
+                        array(
+                            "id" => $id
+                        )
+                    );
 
-                    $this->load->library("upload", $config);
-                    $upload = $this->upload->do_upload("img_url");
-
-                    if($upload){
-                        $uploaded_file = $this->upload->data("file_name");
-                        $data = array(
-                            "url"         => converToSEO($this->input->post("title")),
-                            "title"       => $this->input->post("title"),
-                            "description" => $this->input->post("description"),
-                            "news_type"   => $news_type,
-                            "img_url"     => $uploaded_file,
-                            "video_url"     => '#'
+                    if (!$image){
+                        $alert = array(
+                            "title" => "İşlem başarısız",
+                            "text"  => "Böyle bir haber bulanamadı.",
+                            "type"  => "error"
                         );
+
+                        $this->session->set_flashdata("alert", $alert);
+                        redirect(base_url("news"));
                     }
                     else{
-                        $alert = array(
-                            "title"   => "Opppss",
-                            "text"    => "Resim yüklenme esnasında bir problem oluştu.",
-                            "type"    => "error"
-                        );
-                        $this->session->set_flashdata("alert", $alert);
-                        redirect(base_url("news/update_news/$id"));
+                        $path = "uploads/$this->viewFolder/$image->img_url";
+                        $delete_folder = unlink($path);
+                        if(!$delete_folder) {
+                            $alert = array(
+                                "title" => "İşlem başarısız",
+                                "text" => "Dosya yolu doğru değil veya böyle bir resim yok",
+                                "type" => "error"
+                            );
+
+                            $this->session->set_flashdata("alert", $alert);
+                            redirect(base_url("news"));
+
+                        }
+                        else{
+                            $file_name = converToSEO(pathinfo($_FILES["img_url"]["name"], PATHINFO_FILENAME)). "." . pathinfo($_FILES["img_url"]["name"], PATHINFO_EXTENSION);
+                            $config["allowed_types"] = "jpg|jpeg|png";
+                            $config["upload_path"] = "uploads/$this->viewFolder/";
+                            $config["file_name"] = $file_name;
+
+                            $this->load->library("upload", $config);
+                            $upload = $this->upload->do_upload("img_url");
+
+                            if($upload){
+                                $uploaded_file = $this->upload->data("file_name");
+                                $data = array(
+                                    "url"         => converToSEO($this->input->post("title")),
+                                    "title"       => $this->input->post("title"),
+                                    "description" => $this->input->post("description"),
+                                    "news_type"   => $news_type,
+                                    "img_url"     => $uploaded_file,
+                                    "video_url"     => '#'
+                                );
+                            }
+                            else{
+                                $alert = array(
+                                    "title"   => "Opppss",
+                                    "text"    => "Resim yüklenme esnasında bir problem oluştu.",
+                                    "type"    => "error"
+                                );
+                                $this->session->set_flashdata("alert", $alert);
+                                redirect(base_url("news/update_news/$id"));
+                            }
+                        }
                     }
                 }
                 else{
@@ -373,31 +337,79 @@ class News extends CI_Controller
     }
 
     public function delete($id){
-        
-        $delete = $this->news_model->delete(
+
+        $news = $this->news_model->get(
           array(
               "id" => $id
           )
         );
+        if ($news->news_type == "image"){
+            if ($news){
+                $path = "uploads/$this->viewFolder/$news->img_url";
+                $delete_folder = unlink($path);
+                if(!$delete_folder) {
+                    $alert = array(
+                        "title" => "İşlem başarısız",
+                        "text" => "Dosya yolu doğru değil veya böyle bir resim yok",
+                        "type" => "error"
+                    );
 
-        if($delete){
-            $alert = array(
-                "title"   => "Tebrikler",
-                "text"    => "İşleminiz başarılı bir şekilde gerçekleştirildi.",
-                "type"    => "success"
+                    $this->session->set_flashdata("alert", $alert);
+                    redirect(base_url("news"));
+                }else{
+                    $delete = $this->news_model->delete(
+                        array(
+                            "id" => $id
+                        )
+                    );
+
+                    if($delete){
+                        $alert = array(
+                            "title"   => "Tebrikler",
+                            "text"    => "İşleminiz başarılı bir şekilde gerçekleştirildi.",
+                            "type"    => "success"
+                        );
+
+                    }
+                    else{
+                        $alert = array(
+                            "title"   => "İşlem başarısız",
+                            "text"    => "Lütfen zorunlu olan alanları doldurunuz!",
+                            "type"    => "error"
+                        );
+                    }
+
+                    $this->session->set_flashdata("alert", $alert);
+                    redirect(base_url("news"));
+                }
+            }
+        }
+        else if($news->news_type == "video"){
+            $delete = $this->news_model->delete(
+                array(
+                    "id" => $id
+                )
             );
 
-        }
-        else{
-            $alert = array(
-                "title"   => "İşlem başarısız",
-                "text"    => "Lütfen zorunlu olan alanları doldurunuz!",
-                "type"    => "error"
-            );
-        }
+            if($delete){
+                $alert = array(
+                    "title"   => "Tebrikler",
+                    "text"    => "İşleminiz başarılı bir şekilde gerçekleştirildi.",
+                    "type"    => "success"
+                );
 
-        $this->session->set_flashdata("alert", $alert);
-        redirect(base_url("news"));
+            }
+            else{
+                $alert = array(
+                    "title"   => "İşlem başarısız",
+                    "text"    => "Lütfen zorunlu olan alanları doldurunuz!",
+                    "type"    => "error"
+                );
+            }
+
+            $this->session->set_flashdata("alert", $alert);
+            redirect(base_url("news"));
+        }
 
     }
